WIP: C implementation

+++ main.c
@@ -3,6 +3,7 @@
 #include <sys/types.h>
 #include <string.h>
 #include <stdlib.h>
+#include <unistd.h>
 
 
 char temp[1024];
@@ -18,30 +19,13 @@ int FileExists(const char* path) {
         return (stat(path, &info) == 0) && S_ISREG(info.st_mode);
 }
 
-char* FileToCArray(const char* path) {
-        FILE* f = fopen(path, "rb");
-        if (!f) return NULL;
-
-        if (fseek(f, 0, SEEK_END) != 0) return NULL;
-        long target_file_size = ftell(f);
-        if (fseek(f, 0, SEEK_SET) != 0) return NULL;
-
-        char* file_as_c_array = malloc(256 + (target_file_size * 5));
-        if (file_as_c_array == NULL) return NULL;
-
-        strcpy(file_as_c_array, "const unsigned char DATA[] = {");
-        int c;
-        while ((c = fgetc(f)) != EOF) {
-                sprintf(
-                        file_as_c_array + strlen(file_as_c_array),
-                        "0x%02X,",
-                        (unsigned char)c
-                );
+#define XOR_HASH_LEN 64
+static inline unsigned char* XORHash(const unsigned char* data, size_t data_len) {
+        unsigned char* hash = calloc(XOR_HASH_LEN, 1);
+        for (size_t i = 0; i < data_len; i++) {
+                hash[(i*3) & XOR_HASH_LEN-1] ^= data[i];
         }
-        fclose(f);
-        strcat(file_as_c_array, "}");
-
-        return file_as_c_array;
+        return hash;
 }
 
 int main(int argc, const char* argv[]) {
@@ -67,30 +51,51 @@ int main(int argc, const char* argv[]) {
         strcat(temp, argv[1]);
         system(temp);
 
-        char* file_data_array = FileToCArray("_TEMP.tar");
+        FILE* target_file = fopen("_TEMP.tar", "rb");
+        if (!target_file) {
+                puts("ERROR: Failed to open _TEMP.tar");
+                return 1;
+        }
+
+        if (fseek(target_file, 0, SEEK_END) != 0) {
+                puts("ERROR: Failed to seek to end of _TEMP.tar");
+                return 1;
+        }
+        long target_file_size = ftell(target_file);
+        if (fseek(target_file, 0, SEEK_SET) != 0) {
+                puts("ERROR: Failed to seek to beginning of _TEMP.tar");
+                return 1;
+        }
+
+        char* target_file_data = malloc(target_file_size);
+        if (target_file_data == NULL) {
+                puts("ERROR: Failed to allocate memory for target file data");
+                return 1;
+        }
+
+        if (fread(target_file_data, 1, target_file_size, target_file) < target_file_size) {
+                puts("ERROR: Failed to read data from _TEMP.tar");
+                return 1;
+        }
 
         system("rm _TEMP.tar");
 
-        FILE* f = fopen("_TEMP.c", "w");
-        const unsigned char INCLUDES[] =
-                "#include <stdlib.h>\n"
-                "#include <sys/stat.h>\n"
-                "#include <sys/types.h>\n"
-                "#include <unistd.h>\n"
-                "#include <stdio.h>\n"
-                "#include <stdlib.h>\n";
-        fwrite(INCLUDES, 1, sizeof(INCLUDES), f);
+        unsigned char* _target_file_data_hash = XORHash(target_file_data, target_file_size);
+        unsigned char* target_file_data_hash_hex = calloc(XOR_HASH_LEN*2 + 1, 1);
+        for (int i = 0; i < XOR_HASH_LEN; i++) {
+                sprintf(
+                        target_file_data_hash_hex + i*2,
+                        "%02x",
+                        _target_file_data_hash[i]
+                );
+        }
+        free(_target_file_data_hash);
+
+        puts(target_file_data_hash_hex); // TEMP; TEST
+
         // TODO
-        fclose(f);
 
-        system(
-                "cc -O2 _TEMP.c -o OUT"
-                #ifdef _WIN32
-                        ".exe"
-                #endif
-        );
-
-        system("rm _TEMP.c");
+        free(target_file_data_hash_hex);
 
         return 0;
 }
